+++ build_web.ts
@@ -1,7 +1,7 @@
 import { copy } from "https://deno.land/std@0.165.0/fs/copy.ts";
 
 import sass from "https://deno.land/x/denosass@1.0.4/mod.ts";
-import { bundleFolder } from "./plugos/asset_bundle/builder.ts";
+import { bundleFolder } from "./lib/asset_bundle/builder.ts";
 
 import * as flags from "https://deno.land/std@0.165.0/flags/mod.ts";
 import { patchDenoLibJS } from "./plugos/compile.ts";
