+++ cmd/plug_run.ts
@@ -3,7 +3,7 @@ import { path } from "$common/deps.ts";
 import assets from "../dist/plug_asset_bundle.json" assert {
   type: "json",
 };
-import { AssetBundle } from "../plugos/asset_bundle/bundle.ts";
+import { AssetBundle } from "$lib/asset_bundle/bundle.ts";
 import { determineDatabaseBackend } from "../server/db_backend.ts";
 
 export async function plugRunCommand(
