+++ common/manifest.ts
@@ -6,7 +6,7 @@ import { SlashCommandHookT } from "../web/hooks/slash_command.ts";
 import { PlugNamespaceHookT } from "./hooks/plug_namespace.ts";
 import { CodeWidgetT } from "../web/hooks/code_widget.ts";
 import { MQHookT } from "../plugos/hooks/mq.ts";
-import { EndpointHookT } from "../plugos/hooks/endpoint.ts";
+import { EndpointHookT } from "../server/hooks/endpoint.ts";
 import { PanelWidgetT } from "../web/hooks/panel_widget.ts";
 
 /** Silverbullet hooks give plugs access to silverbullet core systems.
