+++ common/spaces/asset_bundle_space_primitives.ts
@@ -1,5 +1,5 @@
 import { SpacePrimitives } from "./space_primitives.ts";
-import { AssetBundle } from "../../plugos/asset_bundle/bundle.ts";
+import { AssetBundle } from "../../lib/asset_bundle/bundle.ts";
 import { FileMeta } from "../../type/types.ts";
 
 export class AssetBundlePlugSpacePrimitives implements SpacePrimitives {
