--- plug-api/plugos-syscall/mod.ts
@@ -1,7 +1,0 @@
-export * as asset from "./asset.ts";
-export * as events from "./event.ts";
-export * as shell from "./shell.ts";
-export * as YAML from "./yaml.ts";
-export * as mq from "./mq.ts";
-export * from "./syscall.ts";
-export * as datastore from "./datastore.ts";
