--- plug-api/plugos-syscall/syscall.ts
@@ -1,5 +1,0 @@
-declare global {
-  function syscall(name: string, ...args: any[]): Promise<any>;
-}
-
-export const syscall = globalThis.syscall;
