--- plug-api/silverbullet-syscall/mod.ts
@@ -1,10 +1,0 @@
-export * as editor from "./editor.ts";
-export * as markdown from "./markdown.ts";
-export * as space from "./space.ts";
-export * as system from "./system.ts";
-export * as clientStore from "./clientStore.ts";
-export * as sync from "./sync.ts";
-export * as debug from "./debug.ts";
-export * as language from "./language.ts";
-export * as template from "./template.ts";
-export * as codeWidget from "./code_widget.ts";
