+++ plug-api/syscalls.ts
@@ -1,2 +1,18 @@
-export * from "$sb/silverbullet-syscall/mod.ts";
-export * from "$sb/plugos-syscall/mod.ts";
+export * as editor from "./syscalls/editor.ts";
+export * as markdown from "./syscalls/markdown.ts";
+export * as space from "./syscalls/space.ts";
+export * as system from "./syscalls/system.ts";
+export * as clientStore from "./syscalls/clientStore.ts";
+export * as sync from "./syscalls/sync.ts";
+export * as debug from "./syscalls/debug.ts";
+export * as language from "./syscalls/language.ts";
+export * as template from "./syscalls/template.ts";
+export * as codeWidget from "./syscalls/code_widget.ts";
+
+export * as asset from "./syscalls/asset.ts";
+export * as events from "./syscalls/event.ts";
+export * as shell from "./syscalls/shell.ts";
+export * as YAML from "./syscalls/yaml.ts";
+export * as mq from "./syscalls/mq.ts";
+export * from "./syscall.ts";
+export * as datastore from "./syscalls/datastore.ts";
