+++ plug-api/syscalls/asset.ts
@@ -1,5 +1,5 @@
-import { base64DecodeDataUrl } from "../../lib/crypto.ts";
-import { syscall } from "./syscall.ts";
+import { base64DecodeDataUrl } from "$lib/crypto.ts";
+import { syscall } from "../syscall.ts";
 
 export async function readAsset(
   plugName: string,
