+++ plug-api/syscalls/debug.ts
@@ -1,4 +1,4 @@
-import { syscall } from "./syscall.ts";
+import { syscall } from "../syscall.ts";
 
 export function resetClient() {
   return syscall("debug.resetClient");
