+++ plug-api/syscalls/editor.ts
@@ -1,7 +1,7 @@
 import type { FilterOption } from "../../web/types.ts";
-import { PageRef } from "../lib/page_ref.ts";
-import { UploadFile } from "../../type/types.ts";
-import { syscall } from "./syscall.ts";
+import { UploadFile } from "$type/types.ts";
+import { syscall } from "../syscall.ts";
+import type { PageRef } from "$sb/lib/page_ref.ts";
 
 export function getCurrentPage(): Promise<string> {
   return syscall("editor.getCurrentPage");
