+++ plug-api/syscalls/system.ts
@@ -1,6 +1,6 @@
 import type { CommandDef } from "$common/hooks/command.ts";
-import { SyscallMeta } from "../../type/types.ts";
-import { syscall } from "./syscall.ts";
+import { SyscallMeta } from "$type/types.ts";
+import { syscall } from "../syscall.ts";
 
 export function invokeFunction(
   name: string,
