+++ plug-api/syscalls/yaml.ts
@@ -1,4 +1,4 @@
-import { syscall } from "./syscall.ts";
+import { syscall } from "../syscall.ts";
 
 export function parse(
   text: string,
