+++ plugs/editor/upload.ts
@@ -1,5 +1,5 @@
-import { editor, space } from "$sb/silverbullet-syscall/mod.ts";
-import { UploadFile } from "../../type/types.ts";
+import { editor, space } from "$sb/syscalls.ts";
+import { UploadFile } from "$type/types.ts";
 
 const maximumAttachmentSize = 1024 * 1024 * 10; // 10MB
